+++ OpenNos.Data/FamilyDTO.cs
@@ -44,7 +44,7 @@ namespace OpenNos.Data
 
         public bool ManagerCanShout { get; set; }
 
-        public byte MaxSize { get; set; }
+        public short MaxSize { get; set; }
 
         public FamilyAuthorityType MemberAuthorityType { get; set; }
